\name{isFALSE}
\alias{isFALSE}

\title{Check for FALSE}
\description{
  Shortcut for \code{identical(x,FALSE)}
}
\usage{
isFALSE(x)
}

\arguments{
  \item{x}{Any object}
}

\value{
  Returns \code{TRUE} if and only if \code{x} is \code{FALSE}, \code{FALSE}
  otherwise.
}
\author{Andreas Borg}

\seealso{\code{\link{identical}},\code{\link{isTRUE}}}

\keyword{misc}
