\name{RLdata500}
\docType{data}
\alias{RLdata500}
\title{Test data for Record Linkage}
\description{
  This table contains 500 records with artificial personal data of for the 
  evaluation of Record Linkage procedures. 
}
\usage{RLdata500}
\format{A character matrix. Each row represents one record, with the following
  columns:
  \describe{
    \item{fname_c1}{First name, first component}
    \item{fname_c2}{First name, second component}
    \item{lname_c1}{Last name, first component}
    \item{lname_c2}{Last name, second component}
    \item{by}{Year of birth}
    \item{bm}{Month of birth}
    \item{bd}{Day of birth}
  }
}
\source{Generated with the data generation component of Febrl (Freely 
  Extensible Biomedical Record Linkage), version 0.3. See 
  http://datamining.anu.edu.au/projects/linkage.html for details.}
\keyword{datasets}
